Move TestCase one namespace level up
