+++ contribution/generator/src/TrackData/TestCase.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\TrackData\CanonicalData;
+namespace App\TrackData;
 
 class TestCase
 {
